Memory allocators: Revert Q_MALLOC to 8-byte alignment
Commit 1f55d05c2b actually upped Q_MALLOC's address alignment from
8-bytes to 16-bytes.  While this isn't necessarily a bug (in fact,
it makes OpenSIPS more crash-safe), it enables new scenarios where
buffer overrun crashes would happen with F_MALLOC but not with Q_MALLOC.

And since Q_MALLOC is the intended crash debugger, these scenarios
must not be allowed, so we stick to 8-byte alignment.  If anyone needs
16-byte aligned memory, they can just re-define ROUNDTO and re-compile.

+++ mem/f_malloc.c
@@ -178,8 +178,9 @@ struct fm_block *fm_malloc_init(char *address, unsigned long size, char *name)
 
 	/* make address and size multiple of 8*/
 	start=(char*)ROUNDUP((unsigned long) address);
-	LM_DBG("F_OPTIMIZE=%lu, /ROUNDTO=%lu, %zu-bytes aligned\n",
-			F_MALLOC_OPTIMIZE, F_MALLOC_OPTIMIZE/ROUNDTO, ROUNDTO);
+	LM_DBG("F_OPTIMIZE=%lu, /ROUNDTO=%lu, %lu-bytes aligned\n",
+			F_MALLOC_OPTIMIZE, F_MALLOC_OPTIMIZE/ROUNDTO,
+			(unsigned long)ROUNDTO);
 	LM_DBG("F_HASH_SIZE=%lu, fm_block size=%zu, frag_size=%zu\n",
 			F_HASH_SIZE, sizeof(struct fm_block), sizeof(struct fm_frag));
 	LM_DBG("params (%p, %lu), start=%p\n", address, size, start);
